🟡 Add 3871. Count Commas in Range II [Medium] (Java)

+++ Medium/3871-Count-Commas-in-Range-II/solution.java
@@ -1,10 +1,10 @@
 // ═══════════════════════════════════════════════════════
 //  Problem  : 3871. Count Commas in Range II
-//  URL      : https://leetcode.com/problems/count-commas-in-range-ii/?envType=daily-question&envId=2026-09-09
+//  URL      : https://leetcode.com/problems/count-commas-in-range-ii/submissions/2136196230/?envType=daily-question&envId=2026-09-09
 //  Difficulty : Medium
 //  Language : Java
-//  Runtime  : 0 ms
-//  Memory   : 41.7 MB
+//  Runtime  : 1 ms
+//  Memory   : 42.7 MB
 //  Solved   : September 9, 2026
 // ═══════════════════════════════════════════════════════
 
@@ -13,7 +13,7 @@ class Solution {
     
        long ans = 0; 
 
-       for (int i = 1000 ; i <= n ;){
+       for (long i = 1000 ; i <= n ;){
             ans+= (n-i+1);
             
             if(i> Long.MAX_VALUE/1000) break;
